requirement.txt
alembic==1.7.4
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.1
SQLAlchemy==1.4.26
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
